fix or suppress warnings in 2018.2
